+++ webscreen/notification.h
@@ -3,6 +3,6 @@
 
 #include "lvgl.h"
 
-extern const lv_img_dsc_t notification;
+extern const lv_image_dsc_t notification;
 
 #endif  // NOTIFICATION_H